### 2024-02-15 #6356 @editorialbot
[REVIEW]: HofstadterTools: A Python package for analyzing the Hofstadter model
**Submitting author:** @bartandrews (<a href="http://orcid.org/0000-0002-9079-7433">Bartholomew Andrews</a>)
**Repository:** https://github.com/HofstadterTools/HofstadterTools
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.3
