### 2024-02-15 #6356 @AlexBuccheri
I've nearly finished reviewing the JOSS points. In the meantime, I'll provide feedback on what I've tried
