### 2024-02-15 #6356 @AlexBuccheri
Installation
-----------------

The python minimum version isn't explicitly specified in the README, although I can see it's set to 3.10.13 in the `environment.yml.` I would typically expect to see this in a `setup.py` or `pyproject.toml` (the latter being prefered).

## Installation with Conda

There are `environment.yml`s in the project root and the `code/` directories. Diff gives me a subtle difference w.r.t. a few package versions. I assume one is redundant?

From root:

```shell
conda env update --file environment.yml

LibMambaUnsatisfiableError: Encountered problems while solving:
  - nothing provides requested _libgcc_mutex ==0.1 main
  - nothing provides requested _openmp_mutex ==5.1 1_gnu
.
.
.
```

Same issue if I follow the instructions:

```shell
conda env create -f environment.yml
```

This could be because of the specific pinning, and the fact that I'm trying to install deps on `osx-arm64`. If you specify the base OS so I can test this in docker.

- [ ] Remove(?) duplicate `environment.yml` file
- [ ] Note the OS and @AlexBuccheri will retry installation from the `environment.yml` in docker
- [ ] Revise `environment.yml`  so it's more robust

## Installation with Pip

```shell
pip install -r requirements.txt

ERROR: Could not find a version that satisfies the requirement python (from versions: none)
ERROR: No matching distribution found for python
```

I'm not sure it makes sense to specify `python` with no version as a dependency in `requirements.txt`
Removing `python` from the `requirements.txt` allows me to install the deps with pip, and run the example band structure.

- [ ] Consider removing `python` from `requirements.txt`

My suggestion would be to replace the `requirements.txt` with a `pyproject.yml`. Something like:

```yaml
# Note, if not specified setuptools is defaulted to
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "HofstaderTools"
authors = [{name = "Bartholomew Andrews"}]
readme = "README.rst"
requires-python = ">=3.10"
# Pinned project dependencies
# requirements.txt generated from here
dependencies = [
  "jupyter>=1.0.0",
  "matplotlib==3.8.0",
  "numpy==1.26.1",
  ADD MORE
]

# TO ADD
[project.urls]

[tool.setuptools.packages.find]
where = ["code"]

# Pinned project dependencies for development
# requirements-dev.txt generated from here
[project.optional-dependencies]
dev = [
  "pytest>=7.2.0",
  "pytest-cov>=4.0.0",
  "sphinx>=7.2.6",
  "sphinx-rtd-theme==2.0.0",
]
```

- [ ] Consider replacing `requirements.txt` with `pyproject.yml`

which also makes the project installable. I also notice that the CI is explicitly specifying pip packages instead of using `pip install -r requirements.txt` to install. I would change this to remove duplication, and keep everything in sync.

- [ ] Update CI to use `requirements.txt` or `pyproject.yml`, rather than explicitly restate each python dependency
