### 2024-02-15 #6356 @AlexBuccheri
## Running the Tests

One has to set the `$PYTHONPATH` for the package to be found. This could be avoided if the package is made installable with the `pyproject.toml`.

```shell
export PYTHONPATH=$PYTHONPATH:~/HofstadterTools/code
```
assumes an installation location. I'd add a comment or make this pseudocode:

```shell
# Please prepend the path accordingly:
export PYTHONPATH=$PYTHONPATH:path/to/HofstadterTools/code
```

(as I stupidly just copy-paste on first attempt at running the tests).

- [ ] Add comment to `$PYTHONPATH` snippet (or make project installable)

Trivial point: Tests depend on files using relative paths. This prevents me running pytest in `code/`.
This could be made more robust by defining the absolute path. For example:

```python
test_dir = os.path.abspath(os.path.dirname("__file__"))
filename = os.path.join(test_dir, 'butterly', 'butterfly_triangular_q_97_t_1.npz')
```

- [ ] Consider making reference file paths absolute in the tests
