### 2024-02-15 #6356 @AlexBuccheri
# Building the Documentation

Documentation deployment is automated to Github pages, which is really nice. Documentation is compact and thorough. An explicit link to the documentation in the README would be benefical, and/or a note on how to build sphinx locally.

```shell
cd docs/
sphinx-apidoc -o source ../code
make html
```

worked for me.

- [ ] Link Github pages explicitly in README
- [ ] Consider adding instructions for building docs locally
