### 2024-02-15 #6356 @AlexBuccheri
## Design
JOSS doesn't specify to comment on this, but whilst checking the code documentation I had a look at the Hofstadter class. This could be simplified by making Hofstadter class a base class that doesn't implement `unit_cell`, and implements `hamiltonian` method with the default definition (hence allowing to overload where appropriate). Each lattice can then be a child class of Hofstadter:

```python
class Square(Hofstadter)
      def __init__(self, p, q, a0=1, t=None, alpha=1, theta=(1, 3), period=1):
      	  super.__init__(p, q, a0, t, alpha, theta, period)

      def unit_cell(self):
      	    # lattice vectors
            a1 = self.a0 * np.array([1, 0])
            a2 = self.a0 * np.array([0, 1])
            avec_val = np.vstack((a1, a2))
            # basis vectors
            abasis1 = self.a0 * np.array([0, 0])
            abasisvec_val = np.array([abasis1])
            # lattice vectors (MUC)
            aMUC1 = a1
            aMUC2 = self.q * a2
            aMUCvec_val = np.vstack((aMUC1, aMUC2))
            # reciprocal lattice vectors (MUC)
            bMUCvec_val = fm.reciprocal_vectors(aMUCvec_val)
            # symmetry points
            GA = np.array([0, 0])
            Y = np.array([0, 0.5])
            S = np.array([0.5, 0.5])
            X = np.array([0.5, 0])
            num_bands_val = len(abasisvec_val) * self.q
       	    return num_bands_val, avec_val, abasisvec_val, bMUCvec_val, sym_points_val

	def hamiltonian(self, k_val):
	    return fm.BasicSquareHamiltonian(self.t, self.p, self.q, k_val, self.period)

        ...
	etc
```

Or one could follow a factory patten.
