### 2024-02-15 #6356 @AlexBuccheri
I've finished reviewing. The paper's well-written and the code is compact, with good documentation on usage. Once the points outlined here are addressed (as well as the one trivial issue I opened in the repo), I'm happy to approve the paper 🥳
