### 2024-02-16 #6356 @AlexBuccheri
I tried the conda install in docker for Ubuntu focal with Python 3.10 and had the same problem (using native ARM instead of emulating X86).  I also tried dropping extensions of the form `=py310h06a4308_0`, but reconciling the env took forever/was hanging.

I'm not sure if the problem stems from pinning python version, if the pinned versions are hardware + OS specific, or both. However, this does imply that the pinning is too specific to make this portable. If you'd like to retain the environment.yml as a general installation method, perhaps you can regenerate the file with only the package names and versions, not the build strings.

```shell
conda env export --from-history > environment.yml
```

removes the build strings (but also annoyingly drops most of the pinning).

@RMeli Do you have experience with creating portable conda environment files?
