### 2024-02-16 #6356 @RMeli
@AlexBuccheri thank you for looking into this. I doubt such file is portable. When I use `conda list --explicit` to export my environment, usually the platform is also included (`linux-64`, `osx-arm64`, `noarch`, ...). I think it would work only if all the packages are marked as `noarch`, otherwise the dependencies are hardware and OS specific. Did it work if you use x86 emulation? @bartandrews which command did you use to export your environment?

If the software is supposed to work on macOS/Windows I think a better approach is to have a conda environment listing only the required dependencies (and eventually version pins, where needed). Otherwise, clearly state that the package only works on Linux.

Even better, of course, as suggested in https://github.com/openjournals/joss-reviews/issues/6356#issuecomment-1947166385, would be to properly package the software so that `pip` will automatically take care of dependencies. The supported platforms can also be listed in `setup.py`/`pyproject.toml`.
