### 2024-02-16 #6356 @bartandrews
@RMeli My mistake, I believe that I used the command `conda env export > environment.yml` without the `--from-history` flag. I will fix this, and I will also investigate the better options that you suggested for package installation. Thank you again for catching this.
