+++ code/environment.yml
@@ -1,257 +1,0 @@
-name: HT
-channels:
-  - conda-forge
-  - anaconda
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - aiofiles=22.1.0=py310h06a4308_0
-  - aiosqlite=0.18.0=py310h06a4308_0
-  - alabaster=0.7.12=pyhd3eb1b0_0
-  - anyio=3.5.0=py310h06a4308_0
-  - argon2-cffi=21.3.0=pyhd3eb1b0_0
-  - argon2-cffi-bindings=21.2.0=py310h7f8727e_0
-  - asttokens=2.0.5=pyhd3eb1b0_0
-  - atk-1.0=2.36.0=ha1a6a79_0
-  - attrs=23.1.0=py310h06a4308_0
-  - babel=2.11.0=py310h06a4308_0
-  - backcall=0.2.0=pyhd3eb1b0_0
-  - beautifulsoup4=4.12.2=py310h06a4308_0
-  - blas=1.0=mkl
-  - bleach=4.1.0=pyhd3eb1b0_0
-  - brotli=1.0.9=h5eee18b_7
-  - brotli-bin=1.0.9=h5eee18b_7
-  - brotli-python=1.0.9=py310h6a678d5_7
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2023.08.22=h06a4308_0
-  - cairo=1.16.0=hb05425b_5
-  - certifi=2023.7.22=py310h06a4308_0
-  - cffi=1.15.1=py310h5eee18b_3
-  - charset-normalizer=2.0.4=pyhd3eb1b0_0
-  - colorama=0.4.6=py310h06a4308_0
-  - comm=0.1.2=py310h06a4308_0
-  - cryptography=41.0.3=py310hdda0065_0
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - cyrus-sasl=2.1.28=h52b45da_1
-  - dataclasses=0.8=pyh6d0b6a4_7
-  - dbus=1.13.18=hb2f20db_0
-  - debugpy=1.6.7=py310h6a678d5_0
-  - decorator=5.1.1=pyhd3eb1b0_0
-  - defusedxml=0.7.1=pyhd3eb1b0_0
-  - entrypoints=0.4=py310h06a4308_0
-  - exceptiongroup=1.0.4=py310h06a4308_0
-  - executing=0.8.3=pyhd3eb1b0_0
-  - expat=2.5.0=h6a678d5_0
-  - font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
-  - font-ttf-inconsolata=2.001=hcb22688_0
-  - font-ttf-source-code-pro=2.030=hd3eb1b0_0
-  - font-ttf-ubuntu=0.83=h8b1ccd4_0
-  - fontconfig=2.14.1=h4c34cd2_2
-  - fonts-anaconda=1=h8fa9717_0
-  - fonts-conda-ecosystem=1=hd3eb1b0_0
-  - fonttools=4.25.0=pyhd3eb1b0_0
-  - freetype=2.12.1=h4a9f257_0
-  - fribidi=1.0.10=h7b6447c_0
-  - gdk-pixbuf=2.42.10=h5eee18b_0
-  - giflib=5.2.1=h5eee18b_3
-  - glib=2.69.1=he621ea3_2
-  - gobject-introspection=1.72.0=py310hbb6d50b_2
-  - graphite2=1.3.14=h295c915_1
-  - graphviz=2.50.0=h3cd0ef9_0
-  - gst-plugins-base=1.14.1=h6a678d5_1
[…]
