+++ requirements.txt
@@ -1,4 +1,3 @@
-python
 numpy
 matplotlib
 prettytable
