+++ docs/source/theory/butterfly.rst
@@ -1,7 +1,7 @@
 Butterfly
 =========
 
-The famous butterfly spectrum of the Hofstadter model was first plotted by Douglas Hofsatder in 1976 :cite:`Hofstadter76` and later colored in by Joseph Avron in 2003 :cite:`Avron03`. The beautiful fractal structure is a result of the incommensurability between the MUC area and the irreduicble area occupied by a quantum of flux. In this section, we explain how the Hofstadter butterfly is computed and colored in HofstadterTools.
+The famous butterfly spectrum of the Hofstadter model was first plotted by Douglas Hofstadter in 1976 :cite:`Hofstadter76` and later colored in by Joseph Avron in 2003 :cite:`Avron03`. The beautiful fractal structure is a result of the incommensurability between the MUC area and the irreducible area occupied by a quantum of flux. In this section, we explain how the Hofstadter butterfly is computed and colored in HofstadterTools.
 
 The **butterfly spectrum** is a plot of energy :math:`E` against flux density :math:`n_\phi=p/q`, where :math:`p` and :math:`q` are coprime, and is constructed by sweeping over :math:`p` at fixed :math:`q`. As we have seen in the previous section, the bands in the Hofstadter model generally have a :math:`\mathbf{k}`-dependence. However, as we approach the Landau level limit :math:`q\to\infty`, the band structure approaches the Landau level spectrum, where the bands are perfectly flat. In HofstadterTools, we plot the butterfly spectrum using the :math:`\Gamma`-point of each band at :math:`\mathbf{k}=(0,0)`, with the understanding that this is only representative of the band structure when :math:`q` is large. Since a large :math:`q` needs to be chosen in order to plot a convincing butterfly spectrum in any case, this is a reasonable approximation to make. Alternative methods include, for example, taking an average over :math:`\mathbf{k}` for each band and at each flux density, however this increases the computational cost significantly. Moreover, in addition to choosing a large :math:`q` for the butterfly computations, it is also most efficient to choose a prime number, since the flux density is defined as a coprime fraction :math:`n_\phi=p/q`. If a non-prime :math:`q` is chosen, then you will notice gaps in the x-values of the butterfly, since values of :math:`p` that yield an inadmissible :math:`n_\phi` will be skipped.
 
