### 2024-02-17 #6356 @bartandrews
I have now replaced the `environment.yml` file with the one generated using the `--from-history` flag. This is temporary solution, while I implement a better installation method.
