### 2024-02-17 38248ab
make pytest more robust
