+++ code/tests/test_butterfly.py
@@ -53,14 +53,14 @@ def butterfly(q, t, lat, alpha=1, theta=(1, 3), period=1):
     return nphi_list, E_list
 
 
-def test_square():
+def test_square(test_dir):
     """Benchmark the square Hofstadter butterfly. This butterfly is plotted in many papers, e.g. Fig.4(d) of :cite:`Yilmaz17`."""
 
     # current
     nphi_list, E_list = butterfly(97, [1], "square")
     current = np.array([nphi_list, E_list])
     # reference
-    filename = "code/tests/butterfly/butterfly_square_q_97_t_1.npz"
+    filename = test_dir.joinpath('butterfly', 'butterfly_square_q_97_t_1.npz')
     file_data = np.load(filename, allow_pickle=True)
     data = file_data['data'].item()
     nphi_list_ref = data['nphi_list']
@@ -70,14 +70,14 @@ def test_square():
     assert np.allclose(current, reference)
 
 
-def test_square_2():
+def test_square_2(test_dir):
     """Benchmark the zero-quadratic model butterfly against Fig.2 of :cite:`Bauer22`."""
 
     # current
     nphi_list, E_list = butterfly(97, [1, 0, -0.25], "square")
     current = np.array([nphi_list, E_list])
     # reference
-    filename = "code/tests/butterfly/butterfly_square_q_97_t_1_0_-0.25.npz"
+    filename = test_dir.joinpath('butterfly', 'butterfly_square_q_97_t_1_0_-0.25.npz')
     file_data = np.load(filename, allow_pickle=True)
     data = file_data['data'].item()
     nphi_list_ref = data['nphi_list']
@@ -87,14 +87,14 @@ def test_square_2():
     assert np.allclose(current, reference)
 
 
-def test_triangular():
+def test_triangular(test_dir):
     """Benchmark the triangular Hofstadter butterfly. This butterfly is plotted in many papers, e.g. Fig.4(a) of :cite:`Yilmaz17`."""
 
     # current
     nphi_list, E_list = butterfly(97, [1], "triangular")
     current = np.array([nphi_list, E_list])
     # reference
-    filename = "code/tests/butterfly/butterfly_triangular_q_97_t_1.npz"
+    filename = test_dir.joinpath('butterfly', 'butterfly_triangular_q_97_t_1.npz')
     file_data = np.load(filename, allow_pickle=True)
     data = file_data['data'].item()
     nphi_list_ref = data['nphi_list']
@@ -104,14 +104,14 @@ def test_triangular():
     assert np.allclose(current, reference)
 
 
-def test_triangular_2():
+def test_triangular_2(test_dir):
     """Benchmark the 2nd-NN triangular Hofstadter butterfly. This butterfly is plotted in many papers, e.g. Fig.4 of :cite:`Oh00`."""
 
     # current
     nphi_list, E_list = butterfly(97, [0, 1], "triangular", period=2)
     current = np.array([nphi_list, E_list])
     # reference
-    filename = "code/tests/butterfly/butterfly_triangular_q_97_t_0_1_period_2.npz"
+    filename = test_dir.joinpath('butterfly', 'butterfly_triangular_q_97_t_0_1_period_2.npz')
     file_data = np.load(filename, allow_pickle=True)
     data = file_data['data'].item()
     nphi_list_ref = data['nphi_list']
@@ -121,14 +121,14 @@ def test_triangular_2():
     assert np.allclose(current, reference)
 
 
-def test_bravais():
+def test_bravais(test_dir):
     """Benchmark the Bravais Hofstadter butterfly against Fig.4(b-c) of :cite:`Yilmaz17`."""
 
     # current
[…]
