+++ code/tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from pathlib import Path
+
+
+@pytest.fixture(scope="module")
+def test_dir(request):
+    """Return the directory of the currently running test script."""
+
+    return request.path.parent
