+++ code/tests/conftest.py
@@ -1,5 +1,4 @@
 import pytest
-from pathlib import Path
 
 
 @pytest.fixture(scope="module")
