+++ .github/workflows/docs.yml
@@ -16,7 +16,7 @@ jobs:
       - uses: ts-graphviz/setup-graphviz@v1
       - name: Install dependencies
         run: |
-          pip install sphinx==5.3.0 sphinx-rtd-theme pybtex sphinx-copybutton sphinxcontrib-bibtex==2.5.0 sphinx-autodoc-typehints==1.19.5 nbsphinx ipython pydata-sphinx-theme jupyter jupytext pypandoc numpy matplotlib prettytable tqdm
+          pip install -r requirements.txt
       - name: Sphinx build
         run: |
           sphinx-build docs/source docs/build
