+++ requirements.txt
@@ -1,22 +1,23 @@
 # core functionality
-numpy #
-matplotlib #
-prettytable #
-tqdm #
-pytest
+numpy ##
+matplotlib ##
+prettytable ##
+tqdm ##
+pytest #
 # interactivity
-ipython #
-jupyter #
+ipython ##
+jupyter ##
 # documentation
-sphinx==5.3.0 #
-sphinx-rtd-theme #
-sphinxcontrib-bibtex==2.5.0 #
-sphinx-copybutton #
-pydata-sphinx-theme #
-sphinx-autodoc-typehints==1.19.5 #
-nbsphinx #
+sphinx==5.3.0 ##
+sphinx-rtd-theme ##
+sphinxcontrib-bibtex==2.5.0 ##
+sphinx-copybutton ##
+pydata-sphinx-theme ##
+sphinx-autodoc-typehints==1.19.5 ##
+nbsphinx ##
 graphviz
 pybtex #
-jupytext #
-pypandoc #
-pybtex #
+jupytext ##
+pypandoc ##
+pybtex ##
+myst_parser #
