+++ .github/workflows/pytests.yml
@@ -14,7 +14,7 @@ jobs:
       - uses: ts-graphviz/setup-graphviz@v1
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx_rtd_theme myst_parser pybtex sphinx-copybutton sphinxcontrib-bibtex sphinx-autodoc-typehints nbsphinx ipython pydata-sphinx-theme jupyter jupytext pypandoc numpy matplotlib prettytable tqdm pytest
+          pip install -r requirements.txt
       - name: Set the PYTHONPATH
         run: |
           echo "PYTHONPATH=$PYTHONPATH:code" >> "$GITHUB_ENV"
