### 2024-02-17 6d4f8c7
hand-written environment.yml file
