+++ .github/workflows/pytests.yml
@@ -11,6 +11,8 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v3
+        with:
+          python-version: '3.11'
       - uses: ts-graphviz/setup-graphviz@v1
       - name: Install dependencies
         run: |
