+++ code/band_structure.py
@@ -16,8 +16,7 @@ from models.hofstadter import Hofstadter
 from configuration.band_structure import *
 
 
-if __name__ == '__main__':
-
+def main():
     # parse input arguments
     args = fa.parse_input_arguments("band_structure", "Plot the Hofstadter Band Structure.")
     # general arguments
@@ -69,15 +68,15 @@ if __name__ == '__main__':
             if any(geometry_columns):
                 data['eigenvectors_dkx'] = np.zeros((num_bands, num_bands, samp, samp), dtype=np.complex128)  # complex
                 data['eigenvectors_dky'] = np.zeros((num_bands, num_bands, samp, samp), dtype=np.complex128)  # complex
-                data['Dkx'] = np.dot(np.array([1/(samp-1), 0]), bMUCvec[0])
-                data['Dky'] = np.dot(np.array([0, 1/(samp-1)]), bMUCvec[1])
+                data['Dkx'] = np.dot(np.array([1 / (samp - 1), 0]), bMUCvec[0])
+                data['Dky'] = np.dot(np.array([0, 1 / (samp - 1)]), bMUCvec[1])
             for band in tqdm(range(num_bands), desc="Band Construction", ascii=True):
                 for idx_x in range(samp):
-                    frac_kx = idx_x / (samp-1)
+                    frac_kx = idx_x / (samp - 1)
                     if any(geometry_columns):
-                        frac_kx_dkx = (frac_kx + 1/(1000*(samp-1))) % 1
+                        frac_kx_dkx = (frac_kx + 1 / (1000 * (samp - 1))) % 1
                     for idx_y in range(samp):
-                        frac_ky = idx_y / (samp-1)
+                        frac_ky = idx_y / (samp - 1)
                         k = np.matmul(np.array([frac_kx, frac_ky]), bMUCvec)
                         ham = model.hamiltonian(k)
                         eigvals, eigvecs = np.linalg.eigh(ham)
@@ -85,7 +84,7 @@ if __name__ == '__main__':
                         data['eigenvalues'][band, idx_x, idx_y] = eigvals[idx[band]]
                         data['eigenvectors'][:, band, idx_x, idx_y] = eigvecs[:, idx[band]]
                         if any(geometry_columns):
-                            frac_ky_dky = (frac_ky + 1/(1000*(samp-1))) % 1
+                            frac_ky_dky = (frac_ky + 1 / (1000 * (samp - 1))) % 1
                             k_dkx = np.matmul(np.array([frac_kx_dkx, frac_ky]), bMUCvec)
                             k_dky = np.matmul(np.array([frac_kx, frac_ky_dky]), bMUCvec)
                             ham_dkx = model.hamiltonian(k_dkx)
@@ -181,10 +180,13 @@ if __name__ == '__main__':
                         data['wilson_loops'][band, idx_y] = data['wilson_loops'][band - 1, idx_y]
                 for idx_x in range(samp - 1):
                     if group != band_group[band - 1]:
-                        berry_fluxes[band, idx_x, idx_y] = fbs.berry_curv(data['eigenvectors'], band, idx_x, idx_y, group_size)
+                        berry_fluxes[band, idx_x, idx_y] = fbs.berry_curv(data['eigenvectors'], band, idx_x, idx_y,
+                                                                          group_size)
                         # quantum geometry
                         if any(geometry_columns):
-                            geom_tensor = fbs.geom_tensor(data['eigenvectors'], data['eigenvectors_dkx'], data['eigenvectors_dky'], bMUCvec, band, idx_x, idx_y, group_size)
+                            geom_tensor = fbs.geom_tensor(data['eigenvectors'], data['eigenvectors_dkx'],
[…]
