### 2024-02-19 #6356 @katherineding
I second @AlexBuccheri 's suggestion with regards to installation in https://github.com/openjournals/joss-reviews/issues/6356#issuecomment-1947166385. I am on Ubuntu 22.04 with conda `channel_priority` set to `strict`. When I try `conda env create -f environment.yml`,  the solver is unhappy and fails because
```
LibMambaUnsatisfiableError: Encountered problems while solving:
  - package sphinx-5.3.0-pyhd8ed1ab_0 is excluded by strict repo priority
```
While the command succeeds with `flexible` or `disabled` channel priority, it's probably not ideal to require a channel priority change to have the conda environment install correctly.
