### 2024-02-19 #6356 @katherineding
Another extremely minor nit -- I would suggest removing the three __pycache__ directories still floating about in code/
