### 2024-02-19 #6356 @RMeli
@katherineding given that the error says `sphinx-5.3.0-pyhd8ed1ab_0`, are you trying to use the old `environment.yml` or the new one, which does not contain hashes?
