+++ code/__init__.py
@@ -1,1 +1,0 @@
-"""The source code of HofstadterTools."""
