+++ src/HT/__init__.py
@@ -1,0 +1,1 @@
+"""The HofstadterTools package."""
