+++ src/HT/functions/models.py
@@ -1,5 +1,6 @@
 """Functions for the model classes."""
 
+# --- external imports
 import numpy as np
 import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import axes3d
