### 2024-02-19 #6356 @bartandrews
@katherineding Thank you for letting me know. Considering that not even my hand-written `environment.yml` file was robust, I have decided to drop the `environment.yml` option entirely, and stick with `pyproject.toml`. I now have updated my installation instructions accordingly.
