+++ src/HT/plot/__init__.py
@@ -1,0 +1,1 @@
+"""Plotting scripts for HofstadterTools."""
