### 2024-02-20 #6356 @bartandrews
@AlexBuccheri Thank you again for your comments and review. For some reason, I cannot click on your check-boxes. So I will summarize my responses here:

**Installation** From your comments, I realize that `environment.yml` is not really portable (I have tried both with the `--from-history` flag and hand-written, both did not work as intended). Therefore, I have replaced `environment.yml` and `requirements.txt` files with a `pyproject.toml`, as recommended. This seems to work nicely. One concern that I have is with the package `python3-tk`. This is required in order to have an interactive GUI backend for `matplotlib`, and view the plots (crucial for my package), but I am not sure how to list this a requirement in my `pyproject.toml`. It cannot be installed via pip, it needs to be installed with python, similar to pip itself. I think that some installations of python come bundled with pip and tk as standard, but it's still worrying that for some people it will not flag that this dependency is missing.

I have also changed my CI to read from `pyproject.toml`.

**Tests** I have made my project installable, as above. I have also made the test paths absolute, so that `pytest` will work in the project directory or `src` directory. Note that I have now implemented a src layout, for readability.

**Documentation** An explicit link to the github pages is in the 2nd bullet point of the readme. I use my custom URL https://hostadter.tools One question that I had: I wanted to make my custom URL link directly to my github pages documentation, however the only way that I could think of to do this is to name my repo with the ".github.io" suffix. Do you know if there is a better way?

I have also included instructions for building the docs locally in the `docs` part of the `Directory Structure` section of the readme. I use `make clean html`, instead of your `sphinx-apidoc -o source ../code`. I presume that this equivalent?

**Design** Thank you for these suggestions. The Hofstadter sub-classes would certainly rid me of the ugly `if` statements. I have not implemented anything here yet. I need to think about how best to approach this. I may implement other models (related to the Hofstadter model) in the future, so I need to think about design going forward.

**Distribution** I have a test PyPI distribution ready to go now. I am still testing a few things before pushing to the actual PyPI. In particular, I am looking for ways to solve this `python3-tk` issue...

Some other changes that I have made:

- Changed the table column selector to be via flags
- Reduced from two readme files (sphinx and github) to one readme file that is read by both
- Tested (by incrementally reducing the python version in my conda env) that the program, including docs dependencies, works from python 3.7 -> python 3.12 (latest). It may also work with an older python without the docs dependencies, but I have listed Python>=3.7, for simplicity.

In summary, I have implemented all of your suggestions, apart from the design section (for now), and I am waiting to check some things with the installation before I push to actual PyPI.
