+++ pyproject.toml
@@ -22,9 +22,10 @@ dependencies = [
     "tqdm",
     "pytest",
     "ipython",
-    "jupyter"
+    "jupyter",
+    "PyQt5"
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 authors = [{name = "Bartholomew Andrews", email="bartandrews123@gmail.com"}]
 maintainers = [{name = "Bartholomew Andrews", email="bartandrews123@gmail.com"}]
 description = "A Python package for analyzing the Hofstadter model."
