### 2024-02-20 #6356 @bartandrews
UPDATE: I think that I have fixed my `python-tk` issue. It turns out that I do not need `python-tk`, which gives me the TkAgg backend (in place of non-interactive Agg). Instead I can use the QtAgg backend, which can be installed via pip using `pip install QtPy5`. The full list of interactive backends for matplotlib is here: https://matplotlib.org/stable/users/explain/figure/backends.html#interactive-backends

Therefore, I have now added QtPy5 as a core dependency to my `pyproject.toml`. Interestingly, even though they claim that QtPy5 supports Python>=3.7 (https://pypi.org/project/PyQt5/), for me it only worked for Python>=3.9. So I bumped up my requirements accordingly. Now the `band_structure` program should reliably produce an interactive 3D plot.
