### 2024-02-21 1e4dbdb
fix plot scripts
