+++ src/HT/tests/band_structure/__init__.py
@@ -1,0 +1,1 @@
+"""Reference data for HofstadterTools band_structure tests."""
