+++ src/HT/functions/arguments.py
@@ -29,27 +29,28 @@ def parse_input_arguments(program, description):
     band_structure = parser.add_argument_group("band_structure options")
     butterfly = parser.add_argument_group("butterfly options")
 
-    models = ["Hofstadter"]
-    general.add_argument("-mod", "--model", type=str, default="Hofstadter", choices=models, help="name of model")
-    general.add_argument("-a", type=float, default=1, help="lattice constant")
-    general.add_argument("-t", nargs='+', type=float, default=[1], help="list of hopping amplitudes in ascending order [1NN, 2NN, 3NN, ...]")
-    lattices = ["square", "triangular", "bravais", "honeycomb", "kagome", "custom"]
-    general.add_argument("-lat", "--lattice", type=str, default="bravais", choices=lattices, help="name of lattice")
-    general.add_argument("-alpha", type=float, default=1, help="length of a2 Bravais vector relative to a1 (Bravais lattice anisotropy)")
-    general.add_argument("-theta", nargs=2, type=float, default=[1, 3], help="angle between Bravais basis vectors as a fraction of pi (Bravais lattice obliqueness)")
-    general.add_argument("-input", default=False, action='store_true', help="read hopping parameters from hopping_input.txt file")
-    general.add_argument("-save", default=False, action='store_true', help="save the output data and plots")
-    general.add_argument("-log", default=False, action='store_true', help="save the output logs")
-    general.add_argument("-plt_lat", "--plot_lattice", default=False, action='store_true', help="plot the lattice")
-    general.add_argument("-period", "--periodicity", type=int, default=1, help="factor by which to divide A_UC in the flux density")
-    general.add_argument("-dpi", type=int, default=300, help="dots-per-inch resolution of the saved output image")
-    general.add_argument("-ps", "--point_size", type=float, default=1, help="scale factor by which to scale the default point size")
+    if program in ["plot_band_structure", "plot_butterfly"]:
+        general.add_argument("-load", type=str, required=True, help="filename of data to load")
+    if program in ["band_structure", "butterfly"]:
+        models = ["Hofstadter"]
+        general.add_argument("-mod", "--model", type=str, default="Hofstadter", choices=models, help="name of model")
+        general.add_argument("-a", type=float, default=1, help="lattice constant")
+        general.add_argument("-t", nargs='+', type=float, default=[1], help="list of hopping amplitudes in ascending order [1NN, 2NN, 3NN, ...]")
+        lattices = ["square", "triangular", "bravais", "honeycomb", "kagome", "custom"]
+        general.add_argument("-lat", "--lattice", type=str, default="bravais", choices=lattices, help="name of lattice")
+        general.add_argument("-alpha", type=float, default=1, help="length of a2 Bravais vector relative to a1 (Bravais lattice anisotropy)")
+        general.add_argument("-theta", nargs=2, type=float, default=[1, 3], help="angle between Bravais basis vectors as a fraction of pi (Bravais lattice obliqueness)")
+        general.add_argument("-input", default=False, action='store_true', help="read hopping parameters from hopping_input.txt file")
+        general.add_argument("-period", "--periodicity", type=int, default=1, help="factor by which to divide A_UC in the flux density")
+        general.add_argument("-log", default=False, action='store_true', help="save the output logs")
+    if program in ["band_structure", "butterfly", "plot_band_structure", "plot_butterfly"]:
+        general.add_argument("-save", default=False, action='store_true', help="save the output data and plots")
+        general.add_argument("-plt_lat", "--plot_lattice", default=False, action='store_true', help="plot the lattice")
+        general.add_argument("-dpi", type=int, default=300, help="dots-per-inch resolution of the saved output image")
[…]
