+++ src/HT/tests/butterfly/__init__.py
@@ -1,0 +1,1 @@
+"""Reference data for HofstadterTools butterfly tests."""
