+++ src/HT/configuration/__init__.py
@@ -1,0 +1,1 @@
+"""User-defined configuration files for HofstadterTools."""
