### 2024-02-21 20f0ad7
read version number from pyproject.toml in documentation
