### 2024-02-22 #6356 @bartandrews
@AlexBuccheri @katherineding Following some additional testing, I have now added a PyPI release (https://pypi.org/project/HofstadterTools/) and have updated the readme. I have checked that this works on Ubuntu for venv environments with Python 3.9, 3.10, 3.11, 3.12. I have also checked that it works in conda environments, although conda throws an unusual `libGL error`, which as far as I can tell is an issue with Conda and not HofstadterTools. Please let me know if this is also robust now on your computers. Thanks again!
