### 2024-02-22 9e14fcf
adjust figure widths in documentation
