### 2024-02-22 HofstadterTools/HofstadterTools.github.io#4 @katherineding
band_structure should safeguard against non-coprime nphi
This is beyond the scope of the JOSS review, but I noticed this bug(?) so figured I might as well mention it here.

I was testing out the `band_structure` command and saw that while
`band_structure <...other flags...> -nphi 1 4` gives the correct plots,
`band_structure <...other flags...> -nphi 2 8` does not.

The code should probably raise an error for non-prime p/q, or do a GCD and convert any non-prime inputs to the coprime case.
