### 2024-02-23 2d47349
safeguard against non-coprime nphi in band_structure program
