### 2024-02-23 HofstadterTools/HofstadterTools.github.io#4 @bartandrews
Thank you for catching this! For consistency with the butterfly program, where non-coprime nphi are skipped, I now raise a ValueError when a non-coprime nphi is parsed to the band_structure program.
