+++ .github/workflows/docs_remote.yml
@@ -1,0 +1,33 @@
+name: docs_remote
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  docs_remote:
+    runs-on: ubuntu-20.04
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v3
+        with:
+          python-version: '3.11'
+      - uses: ts-graphviz/setup-graphviz@v1
+      - name: Install dependencies
+        run: |
+          pip install ".[docs]"
+      - name: Sphinx build
+        run: |
+          sphinx-build docs/source docs/build
+      - name: Deploy to external repository
+        uses: cpina/github-action-push-to-another-repository@main
+        env:
+          SSH_DEPLOY_KEY: ${{ secrets.SSH_DEPLOY_KEY }}
+        with:
+          source-directory: 'docs/build/html'
+          destination-github-username: 'HofstadterTools'
+          destination-repository-name: 'HTwebsite.github.io'
+          user-email: bartandrews123@gmail.com
+          target-branch: gh-pages
