+++ .github/workflows/docs_remote.yml
@@ -10,11 +10,11 @@ jobs:
   docs_remote:
     runs-on: ubuntu-20.04
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: '3.11'
-      - uses: ts-graphviz/setup-graphviz@v1
+      - uses: ts-graphviz/setup-graphviz@v2
       - name: Install dependencies
         run: |
           pip install ".[docs]"
