+++ .github/workflows/docs_remote.yml
@@ -26,6 +26,6 @@ jobs:
         env:
           SSH_DEPLOY_KEY: ${{ secrets.SSH_DEPLOY_KEY }}
         with:
-          source-directory: 'docs'
+          source-directory: 'docs/build'
           destination-github-username: 'HofstadterTools'
           destination-repository-name: 'HTwebsite.github.io'
