### 2024-02-23 #6356 @bartandrews
I have found a possible solution to my ".github.io" suffix question. As far as I know, I can only point custom URLs to github repositories with this suffix. However, instead of building the documentation in the same repository into a different branch, I can configure the CI to deploy the built documentation into a different repository (within the HofstadterTools organization). @RMeli Is it okay if I rename the repository from "HofstadterTools.github.io" to "HofstadterTools"? My plan is to have 2 repos in the HofstadterTools organization: "HofstadterTools" containing the project (like it is now), and "HofstadterTools.github.io" containing the built documentation hosted using github pages.
