+++ .github/workflows/docs_remote.yml
@@ -22,6 +22,8 @@ jobs:
         run: |
           sphinx-build docs/source docs/build
       - name: Deploy to external repository
+        run: |
+          echo hofstadter.tools > docs/build/CNAME
         uses: cpina/github-action-push-to-another-repository@main
         env:
           SSH_DEPLOY_KEY: ${{ secrets.SSH_DEPLOY_KEY }}
