+++ .github/workflows/docs_remote.yml
@@ -1,10 +1,9 @@
 name: docs_remote
 
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
+on: [push, pull_request, workflow_dispatch]
+
+permissions:
+  contents: write
 
 jobs:
   docs_remote:
