### 2024-02-23 #6356 @RMeli
Sorry @bartandrews, I didn't catch this issue before. I should inquiry how renaming a repository during review works, I'll let you know.

However, I'm not entirely sure why two repositories are needed. `http(s)://<organization>.github.io` is the top-level website for an organisation. However, documentation can usually live within the repository under `http(s)://<organization>.github.io/<repository>` (i.e. this is a project-specific site.

> An explicit link to the github pages is in the 2nd bullet point of the readme. I use my custom URL https://hostadter.tools/ One question that I had: I wanted to make my custom URL link directly to my github pages documentation, however the only way that I could think of to do this is to name my repo with the ".github.io" suffix.

Looking at [GitHub Documentation](https://docs.github.com/en/pages/configuring-a-custom-domain-for-your-github-pages-site/about-custom-domains-and-github-pages#using-a-custom-domain-across-multiple-repositories), however, it seems that you can override the default project domain by adding a custom domain to the individual repository.

I think the problem here is that you have put your code into [HofstadterTools.github.io](https://github.com/HofstadterTools/HofstadterTools.github.io), which is technically supposed to be the repo for the organisation website. So renaming it to `HofstadterTools` should allow you to have a project-specific website normally under `http(s)://HofstadterTools.github.io/HofstadterTools` (which should be possible to override).
