### 2024-02-23 #6356 @RMeli
I just checked, and there is no problem [changing the repository during review](https://joss.readthedocs.io/en/latest/editorial_bot.html#changing-the-repository), if needed.
