### 2024-02-23 #6356 @bartandrews
Thank you for letting me know! I will think about this... Most probably, I will change it, as you say, because it is non-standard to have the code in "HofstadterTools.github.io". I think "HofstadterTools" makes more sense
