### 2024-02-23 #6356 @RMeli
> it is non-standard to have the code in "HofstadterTools.github.io". I think "HofstadterTools" makes more sense

Yes, I agree. But up to you how you want to proceed here.
