+++ paper/paper.md
@@ -29,7 +29,7 @@ The Hofstadter model successfully describes the behavior of non-interacting quan
 
 # Statement of need
 
-The purpose of HofstadterTools is to consolidate the fragmented theory and code relevant to the Hofstadter model into one well-documented Python package, which can be used easily by non-specialists as a benchmark or springboard for their own research projects. The Hofstadter model [@Harper55; @Azbel64; @Hofstadter76] is an iconic tight-binding model in physics and famously yields a fractal energy spectrum as a function of flux density, as shown in Figs. \ref{fig:square}, \ref{fig:triangular}, \ref{fig:honeycomb}, and \ref{fig:kagome}. Consequently, it is often treated as an add-on to larger numerical packages, such as pyqula [@pyqula], DiagHam [@DiagHam], and TeNPy [@TeNPy], or simply included as supplementary code together with research articles [@Bodesheim23]. However, the Hofstadter model's generalizability, interdisciplinary appeal, and recent experimental realization, motivates us to create a dedicated package that can provide a detailed analysis of its band structure, in the general case.
+The purpose of HofstadterTools is to consolidate the fragmented theory and code relevant to the Hofstadter model into one well-documented Python package, which can be used easily by non-specialists as a benchmark or springboard for their own research projects. The Hofstadter model [@Harper55; @Azbel64; @Hofstadter76] is an iconic tight-binding model in physics and famously yields a fractal energy spectrum as a function of flux density, as shown in Figs. \ref{fig:square}, \ref{fig:triangular}, \ref{fig:honeycomb}, and \ref{fig:kagome}. Consequently, it is often treated as an add-on to larger numerical packages, such as WannierTools [@WannierTools], pyqula [@pyqula], and DiagHam [@DiagHam], or simply included as supplementary code together with research articles [@Bodesheim23]. However, the Hofstadter model's generalizability, interdisciplinary appeal, and recent experimental realization, motivates us to create a dedicated package that can provide a detailed analysis of its band structure, in the general case.
 
 1) **Generalizability.** The Hofstadter model was originally studied in the context of electrons hopping in a periodic potential coupled to a perpendicular magnetic field. However, the model transcends this framework and can be extended in numerous directions. For example, the Peierls phases that arise in the Hamiltonian due to the magnetic field [@Peierls33] can also be generated using artificial gauge fields [@Goldman14] or Floquet modulation [@Eckardt17]. Moreover, the full scope of the Hofstadter model is still being revealed, with papers on its application to hyperbolic lattices [@Stegmaier22], higher-dimensional crystals [@DiColandrea22], and synthesized materials [@Bodesheim23], all published within the last couple of years.
 
