+++ paper/paper.bib
@@ -52,18 +52,16 @@ url = {http://jetp.ras.ru/cgi-bin/e/index/e/19/3/p634?a=list}
 	year = {2001}
 }
 
-@article{TeNPy,
-	title={{Efficient numerical simulations with Tensor Networks: Tensor Network  Python (TeNPy)}},
-	author={J. Hauschild and F. Pollmann},
-	journal={SciPost Physics Lecture Notes},
-	pages={5},
-	year={2018},
-	publisher={SciPost},
-	doi={10.21468/SciPostPhysLectNotes.5},
-	url={https://scipost.org/10.21468/SciPostPhysLectNotes.5},
-	archiveprefix={arXiv},
-	eprint={1805.00055},
-	note={Code available from \url{https://github.com/tenpy/tenpy}},
+@article{WannierTools,
+title = "Wannier{T}ools : {A}n open-source software package for novel topological materials",
+journal = "Computer Physics Communications",
+volume = "224",
+pages = "405-416",
+year = "2018",
+issn = "0010-4655",
+doi = "10.1016/j.cpc.2017.09.033",
+url = "http://www.sciencedirect.com/science/article/pii/S0010465517303442",
+author = "Q. Wu and S. Zhang and H.-F. Song and M. Troyer and A. A. Soluyanov"
 }
 
 @Article{Bodesheim23,
