+++ paper/paper.bib
@@ -53,7 +53,7 @@ url = {http://jetp.ras.ru/cgi-bin/e/index/e/19/3/p634?a=list}
 }
 
 @article{WannierTools,
-title = "Wannier{T}ools : {A}n open-source software package for novel topological materials",
+title = "Wannier{T}ools: {A}n open-source software package for novel topological materials",
 journal = "Computer Physics Communications",
 volume = "224",
 pages = "405-416",
