+++ docs/source/references.bib
@@ -967,3 +967,15 @@ URL = {https://www.science.org/doi/abs/10.1126/science.aao1401}
   doi = {10.1103/PhysRevLett.131.176501},
   url = {https://link.aps.org/doi/10.1103/PhysRevLett.131.176501}
 }
+
+@article{WannierTools,
+title = "Wannier{T}ools: {A}n open-source software package for novel topological materials",
+journal = "Computer Physics Communications",
+volume = "224",
+pages = "405-416",
+year = "2018",
+issn = "0010-4655",
+doi = "10.1016/j.cpc.2017.09.033",
+url = "http://www.sciencedirect.com/science/article/pii/S0010465517303442",
+author = "Q. Wu and S. Zhang and H.-F. Song and M. Troyer and A. A. Soluyanov"
+}
