### 2024-02-26 17bbd4b
Create CODE_OF_CONDUCT.md
