### 2024-02-26 6ed303b
add citation.cff file
