### 2024-02-27 #6356 @AlexBuccheri
@bartandrews

Thanks for the positive response w.r.t. all the feedback. The changes look great! Just to follow up on those last points:

**Installation**

Installation works smoothly for me with python 3.11 on Mac. Tests work out of the box, as does installing docs dependencies and building 🙌🏻

**Documentation**

`make clean html` is equivalent to the command I suggested, and the _better_ choice. W.r.t. documentation would just add this comment after `pip install -e ".[docs]"` in the top-level README:

To build the documentation, run:

```
make clean html
```

in the `docs/` folder.

But this is super-trivial and non-blocking. I'll leave it to your discretion.

W.r.t. the documentation deployment issues you were having, I would not have been helpful as the main project I'm currently associated with runs on Gitlab, with custom buildbot CI (sigh). But it's great that you found a soluton and @RMeli  added some context - I made a note for my own reference 😄

I assume I can run this command:
