### 2024-02-28 #6356 @bartandrews
@AlexBuccheri Thank you again for catching those issues, and for your suggestions for improvement. I'm glad that the package is more robust now and works on Mac OS 🎉 I will update the readme as recommended, when I next get the chance
