### 2024-03-01 #6356 @katherineding
I have finished reviewing the code repository and paper submission. The code is well-documented and clean. This package addresses a clear need in the topological band structure community for an accessible way to perform Hofstadter and Hofstadter-adjacent tight-binding model calculations, particularly with associated quantum geometry and topology information. The revisions suggested by @AlexBuccheri have already addressed the bigger issues relating to installation, packaging, and testing.

I verified that basic installation, advanced installation, running tests, and building local documentation succeeds on my Ubuntu 22.04, Python 3.11 system.

I am delighted to recommend acceptance! :partying_face:

Below, I list some minor suggestions that I think will improve the user experience of this package. They are non-blocking, however, so please feel free to ignore them.

### Packaging

- Do you have plans for releasing HofstadterTools to conda-forge? It seems this is not too hard to do given an existing PyPI release, as per https://conda-forge.org/docs/maintainer/adding_pkgs/. This would provide an easy install route for users whose preferred package manager is conda and don't want to mix it with pip.

### Documentation

- Consider adding back instructions for how to use `venv` with `HofstadterTools`. This would save people a google if they're not familiar with `venv`. (I for one needed to look this up!) If I recall correctly, these instructions were present in an older version of the docs. I would simply add after the sentence "To avoid dependency clashes, we recommend the use of a python virtual environment, such as `venv`." something like:

> To use `venv`, create a virtual environment: `python -m venv /path/to/env` and activate it: `source /path/to/env/bin/activate`. Then you may install `HofstadterTools` into the newly activated virtual environment by running `pip install HofstadterTools`. This environment now would only contain `HofstadterTools` and its dependencies.
>
> `source /path/to/env/bin/activate` must be run each time a new terminal shell is opened before `HofstadterTools` can be used.
>
> To switch virtual environments. Run `deactivate` then source a different desired environment.
>
> To remove HofstadterTools and its dependencies, run `deactivate` and remove the entire `/path/to/env` folder.

- Consider making more explicit (in documentation/help message) the difference in functionality between `band_structure` and `butterfly`. I am guessing that mainly `band_structure` considers one specific field strength (and also produces quantum geometry information), while `butterfly` runs through a while array of field strengths.

- Consider documenting the meaning of the high symmetry points "Y" and "S" in output of `Hofstadter.unit_cell()`. "S" in particular does not seem to be standard notation, please correct me if I'm wrong on this.

- JOSS requires community guidelines for third parties wishing to 2) Report issues or problems with the software and 3) Seek support. Currently the README only has a "Contributing" section where its probably implicit that support requests should go to @bartandrews. It would probably be good to add a few explicit words about reporting issues and seeking support (e.g. if you would prefer people use the github issues/discussions system v.s. directly contact you).
