### 2024-03-01 #6356 @bartandrews
@katherineding Thank you for your suggestions for improvement! :pray: I am grateful for your feedback regarding the installation, the issue that you spotted https://github.com/HofstadterTools/HofstadterTools/issues/4 , and also the conda-forge / documentation advice above. I will implement these changes as soon as possible
