### 2024-03-01 HofstadterTools/HofstadterTools#4 @RMeli
Related to https://github.com/openjournals/joss-reviews/issues/6356
