### 2024-03-04 c59cab6
add conda-forge to readme
