### 2024-03-04 103bc80
improve programs and contributing readme sections
