### 2024-03-07 fb67642
improve reference paths for 2D band structures
