+++ docs/source/theory/band_structure.rst
@@ -6,17 +6,17 @@ The band structure for the Hofstadter model is computed by constructing a set of
 To recap from the previous section, for Landau gauge in the y-direction :math:`\mathbf{A}=-By \hat{\mathbf{e}}_x`, the general formula for the **Peirels phase** acquired by an electron hopping from site :math:`i=(X_i,Y_i)` to site :math:`j=(X_j,Y_j)` is given as
 
 .. math::
-	\theta_{ij} = -2\pi n_\phi \Delta X \left( Y_i + \frac{\Delta Y}{2} \right),
+  \theta_{ij} = -2\pi n_\phi \Delta X \left( Y_i + \frac{\Delta Y}{2} \right),
 
 where :math:`\Delta X = X_j - X_i`, :math:`\Delta Y = Y_j - Y_i`, and we have defined the flux density with respect to the UC area, such that :math:`n_\phi=BA_\mathrm{UC}/\phi_0 = Ba^2/\phi_0`. In what follows, let us index the UC coordinates using :math:`(m,n)`. The **time-independent Schrödinger equation** for the conventional nearest neighbor (NN) square-lattice Hofstadter model is then given as
 
 .. math::
-	E \Psi_{m,n} = -t e^{-\mathrm{i}2\pi n_\phi n} \Psi_{m+1, n} - t \Psi_{m, n+1} + \mathrm{H.c.},
+  E \Psi_{m,n} = -t e^{-\mathrm{i}2\pi n_\phi n} \Psi_{m+1, n} - t \Psi_{m, n+1} + \mathrm{H.c.},
 
 where :math:`\Psi_{m,n}` is the wavefunction at unit cell :math:`(m,n)` and :math:`t` is the hopping amplitude. Since we have taken Landau gauge in the y-direction, it is reasonable to assume plane wave behavior in the x-direction. Hence we can simplify the equation further by invoking the **plane wave ansatz** :math:`\Psi_{m,n} = e^{\mathrm{i}\mathbf{r}\cdot\mathbf{k}}\psi_{n}`, such that
 
 .. math::
-	E \psi_{n} = -t e^{-\mathrm{i}2\pi n_\phi n + \mathrm{i}k_x} \psi_{n} - t e^{\mathrm{i} k_y}\Psi_{m} + \mathrm{H.c.},
+  E \psi_{n} = -t e^{-\mathrm{i}2\pi n_\phi n + \mathrm{i}k_x} \psi_{n} - t e^{\mathrm{i} k_y}\Psi_{m} + \mathrm{H.c.},
 
 which simplifies to
 
@@ -54,6 +54,42 @@ and hence there are 4 energy bands in the spectrum, as shown in the figures belo
     :width: 32 %
     :alt: 2D Band Structure
 
+.. note::
+  It is not possible to define a high-symmetry path for the generalized Hofstadter model as a continuous function of lattice anisotropy and obliqueness. Instead, we use reference paths to plot the 2D band structures.
+
+  .. list-table::
+   :widths: 10 30 30 30
+   :header-rows: 1
+
+   * - Lattice
+     - UC Lattice Vectors
+     - UC Basis Vectors
+     - Reference Path
+   * - ``square``
+     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=a(0, 1)`
+     - :math:`(0, 0)`
+     - :math:`\Gamma\to X \to M \to Y \to \Gamma`
+   * - ``triangular``
+     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=a (1/2, \sqrt{3}/2)`
[…]

[…]
