### 2024-03-07 09944e8
update 2D band structure figure
