+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), 
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## v1.0.0
+
+- Submission to the Journal of Open Source Software (JOSS).
+
+## v1.0.1
+
+- Corrections from first JOSS reviewer. See the [review thread](https://github.com/openjournals/joss-reviews/issues/6356) for further details.
