+++ pyproject.toml
@@ -17,7 +17,7 @@ pythonpath = ["src"]
 
 [project]
 name = "HofstadterTools"
-version = "1.0.1"
+version = "1.0.2"
 dependencies = [
     "numpy",
     "matplotlib",
