### 2024-03-08 #6356 @bartandrews
@katherineding Thank you again for your feedback and suggestions for the code. It is much appreciated! :smiley: I append my responses below:

**Packaging**
- I have now added the package to conda-forge.

**Documentation**
- I have now expanded the instructions for installation and environment management using venv and conda.
- I have summarized the difference between the `band_structure` and `butterfly` programs in the Quick Start section.
- I have added a note in the Theory>Band Structure section of the documentation explaining how the paths are computed. This comment also made me think more generally about the paths in the 2D band structures. Since it is not possible to define a high-symmetry path in the generalized Hofstadter model that is a continuous function of lattice anisotropy and obliqueness, I have 2 options: (1) Write an algorithm to compute a high-symmetry path (if one even exists) on a case-by-case basis, or (2) define a reference path through k-space that will not always be a high-symmetry path. I have decided on (2) because I couldn't think of an intuitive way of implementing (1), and I'm not convinced that the paths from (1) would always be informative, e.g. for lattices with few symmetries. I have now changed the x-axis label on the 2D band structures from "symmetry path" to "reference path" to reflect this, and I have made it clear in the documentation that the reference paths used are for easy comparison and do not always correspond to high-symmetry paths. I am open to any suggestions that you have. NB: I use $\Gamma\to X \to M \to Y \to \Gamma$ instead of the standard $\Gamma \to X \to M \to \Gamma$ for the square lattice because I define these points w.r.t. the MUC reciprocal lattice vectors and the former captures the key properties of the band structure in the presence of a magnetic field. Similarly for the triangular lattice, etc.
- I have added explicit sentences in the Contributing section.
