### 2024-03-11 7358c3f
add missing DOIs
