+++ paper/paper.bib
@@ -248,7 +248,8 @@ eprint = {https://www.worldscientific.com/doi/pdf/10.1142/9781848160224_0014}
       eprint={2310.05758},
       archivePrefix={arXiv},
       primaryClass={cond-mat.str-el},
-      note={arXiv:2310.05758 [cond-mat.str-el]}
+      note={arXiv:2310.05758 [cond-mat.str-el]},
+      doi={10.48550/arXiv.2310.05758}
 }
 
 @article{Andrews20,
@@ -273,7 +274,8 @@ eprint = {https://www.worldscientific.com/doi/pdf/10.1142/9781848160224_0014}
       eprint={2308.10935},
       archivePrefix={arXiv},
       primaryClass={cond-mat.str-el},
-      note={arXiv:2308.10935 [cond-mat.str-el]}
+      note={arXiv:2308.10935 [cond-mat.str-el]},
+      doi={10.48550/arXiv.2308.10935}
 }
 
 @article{Jackson15,
