+++ docs/source/theory/band_structure.rst
@@ -74,15 +74,15 @@ and hence there are 4 energy bands in the spectrum, as shown in the figures belo
      - :math:`(0, 0)`
      - :math:`\Gamma \to K \to M \to K' \to \Gamma`
    * - ``bravais``
-     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=a \alpha (\cos \theta, \sin \theta)`
+     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=\alpha a (\cos \theta, \sin \theta)`
      - :math:`(0, 0)`
      - :math:`\begin{cases}\Gamma \to K \to M \to K' \to \Gamma, \;\;\; &\text{when }\theta\text{ is a multiple of }\pi/3\text{ but not }\pi/2 \\ \Gamma \to X \to M \to Y \to \Gamma, \;\;\; &\text{otherwise}\end{cases}`
    * - ``honeycomb``
-     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=a \alpha (\cos \theta, \sin \theta)`
+     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=\alpha a (\cos \theta, \sin \theta)`
      - :math:`(0, 0)`, :math:`(a_{1,x}/2, a_{2,y}/3)`
      - :math:`\Gamma \to K \to M \to K' \to \Gamma`
    * - ``kagome``
-     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=a \alpha (\cos \theta, \sin \theta)`
+     - :math:`\mathbf{a}_1 = a (1, 0)`, :math:`\mathbf{a}_2=\alpha a (\cos \theta, \sin \theta)`
      - :math:`(0, 0)`, :math:`(a_{1,x}/2, 0)`, :math:`(a_{2,x}/2, a_{2,y}/2)`
      - :math:`\Gamma \to K \to M \to K' \to M' \to \Gamma`
 
@@ -90,7 +90,7 @@ and hence there are 4 energy bands in the spectrum, as shown in the figures belo
 
   We emphasize that the paths defined above are only *high-symmetry paths* in special cases, where the corresponding symmetries are present, such as at zero magnetic field. In all other cases, these are simply *reference paths* through :math:`k`-space to facilitate easy comparison, and the full 3D band structure should also be examined.
 
-In the general case, the procedure follows in a similar way. We start by writing down the time-independent Schrödinger equation for each site in the basis, which we index using lowercase Greek letters :math:`\alpha,\beta`. This yields a set :math:`N_\mathrm{b}` simultaneous equations, which we can write in matrix form, such that :math:`\mathbf{H}\boldsymbol{\psi}=E\boldsymbol{\psi}`, where :math:`\boldsymbol{\psi}=(\psi^0, \psi^1, \dots, \psi^{N_\mathrm{b}-1})^\intercal` is a vector of length :math:`N_\mathrm{b}`, and the :math:`N_\mathrm{b}\times N_\mathrm{b}` Hamiltonian matrix is given as
[…]

[…]
