### 2024-03-11 7362ef6
add hopping_input.txt
