+++ src/HT/configuration/hopping_input.txt
@@ -1,0 +1,2 @@
+1	1
+3	-0.25
