### 2024-03-11 #6356 @RMeli
Paper looks good. The missing DOIs are for two unpublished repositories, and for a very old paper:
> Azbel, M. Y. (1964). Energy spectrum of a conduction electron in a magnetic field. Journal of Experimental and Theoretical Physics, 19(3), 634–645.
