+++ CHANGELOG.md
@@ -16,3 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## v1.0.2
 
 - Corrections from second JOSS reviewer. See the [review thread](https://github.com/openjournals/joss-reviews/issues/6356) for further details.
+
+## v1.0.3
+
+- Accepted version in JOSS. See the [review thread](https://github.com/openjournals/joss-reviews/issues/6356) for further details.
