+++ docs/source/index.rst
@@ -35,7 +35,3 @@ Indices and tables
 * :ref:`genindex`
 * :ref:`modindex`
 * :ref:`search`
-
-.. note::
-
-   This project is under active development.

[…]
