### 2024-03-12 485d26f
update CITATION.cff
