+++ src/HT/functions/arguments.py
@@ -53,7 +53,7 @@ def parse_input_arguments(program, description):
         band_structure.add_argument("-samp", type=int, default=101, help="number of samples in linear direction")
         band_structure.add_argument("-nphi", nargs=2, type=int, default=[1, 4], help="flux density")
         band_structure.add_argument("-bgt", type=float, default=0.01, help="band gap threshold")
-        band_structure.add_argument("-load", default=False, action='store_true', help="load data from file")
+        band_structure.add_argument("-load", type=str, default=False, help="load data from file")
         band_structure.add_argument("-topo", "--topology", default=False, action='store_true', help="print the topology columns")
         band_structure.add_argument("-geom", "--geometry", default=False, action='store_true', help="print the basic quantum geometry columns")
         columns = ["band", "group", "isolated", "width", "gap", "gap_width", "std_B", "C", "std_g", "av_gxx", "std_gxx", "av_gxy", "std_gxy", "T", "D"]
