+++ docs/source/troubleshooting.rst
@@ -94,9 +94,9 @@ The band structure with zero magnetic field can be computed by setting the flux 
 
 .. admonition:: Question
 
-	I have saved a band structure to file and I would now like to regenerate the output table using the ``plot_band_structure`` program. How can I do this?
+	I have saved a band structure to file and I would now like to regenerate the output table. How can I do this?
 
-Currently, the ``plot_band_structure`` program can regenerate the band structure plots and modify their appearance, using the ``band_structure`` function in ``src/HT/functions/plotting.py``. Since the program calls the plotting function directly, it cannot be used to regenerate the output table. However, the output table can be saved to file by running the original ``band_structure`` program with the ``-log`` flag.
+The ``band_structure`` program takes an input file via the ``-load`` argument. By using this flag, the band structure is loaded from file and only the band properties are computed. You can then regenerate and/or reconfigure the output table as usual. Additionally, the output table can be saved to file by running ``band_structure`` with the ``-log`` flag. We note that the ``plot_band_structure`` program can only regenerate the band structure plots and modify their appearance, using the ``band_structure`` function in ``src/HT/functions/plotting.py``. Since ``plot_band_structure`` calls the plotting function directly, it cannot be used to regenerate the output table.
 
 Butterfly Questions
 -------------------
