### 2024-03-12 #6356 @bartandrews
@RMeli Thank you for checking the paper! I have also checked it again, and I am happy with it. (As far as I know, there are no DOIs available for the references marked as "missing DOIs".)

The new version number is v1.0.3 and the DOI is here: https://doi.org/10.5281/zenodo.10809833
