### 2024-03-12 #6356 @RMeli
Thank you @bartandrews for submitting to JOSS. Based on the extensive comments from the reviewers (and my own assessment), I'm happy to recommend this paper for acceptance! The EiC will now perform the final checks. Thank you for engaging so well with the review process.

Thank you very much again @AlexBuccheri and @katherineding for the in depth and constructive review!
