### 2024-03-15 b146e65
add additional references
