+++ docs/source/references.bib
@@ -995,3 +995,44 @@ doi = "10.1016/j.cpc.2017.09.033",
 url = "http://www.sciencedirect.com/science/article/pii/S0010465517303442",
 author = "Q. Wu and S. Zhang and H.-F. Song and M. Troyer and A. A. Soluyanov"
 }
+
+@article{Xiao03,
+  title = {Landau levels in the case of two degenerate coupled bands: {K}agom\'e lattice tight-binding spectrum},
+  author = {Xiao, Y. and Pelletier, V. and Chaikin, P. M. and Huse, D. A.},
+  journal = {Physical Review B},
+  volume = {67},
+  issue = {10},
+  pages = {104505},
+  numpages = {9},
+  year = {2003},
+  month = {Mar},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevB.67.104505},
+  url = {https://link.aps.org/doi/10.1103/PhysRevB.67.104505}
+}
+
+@article{Vidal98,
+  title = {Aharonov-{B}ohm Cages in Two-Dimensional Structures},
+  author = {Vidal, J. and Mosseri, R. and Douçot, B.},
+  journal = {Physical Review Letters},
+  volume = {81},
+  issue = {26},
+  pages = {5888--5891},
+  numpages = {0},
+  year = {1998},
+  month = {Dec},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.81.5888},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.81.5888}
+}
+
+@book{Satija16,
+author = {Satija, I. I.},
+title = {Butterfly in the {Q}uantum {W}orld},
+publisher = {Morgan & Claypool Publishers},
+year = {2016},
+series = {2053-2571},
+isbn = {978-1-6817-4117-8},
+url = {https://dx.doi.org/10.1088/978-1-6817-4117-8},
+doi = {10.1088/978-1-6817-4117-8}
+}
